requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
argcomplete==3.5.3
asgiref==3.8.1
astroid==3.3.8
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cohere==5.14.0
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
durationpy==0.9
eval_type_backport==0.2.2
fastapi==0.115.11
fastavro==1.10.0
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.69.0
griffe==1.6.0
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isort==6.0.1
jiter==0.8.2
jsonpath-python==1.0.6
kubernetes==32.0.1
logfire-api==3.7.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.5.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.65.3
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pipx==1.7.1
platformdirs==4.3.6
posthog==3.18.1
prompt_toolkit==3.0.50
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai @ git+https://github.com/pydantic/pydantic-ai.git@13ece6d669bf37a7a2410567c8f60844d30a37f6
pydantic-ai-slim==0.0.35
pydantic-graph==0.0.35
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pylint==3.3.4
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250301
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
zipp==3.21.0
